fix bug where there is no way to navigate away from the trade history page remove trade tab from ios builds
