+++ libwallet/btc.go
@@ -8,7 +8,6 @@ import (
 	"github.com/btcsuite/btcwallet/waddrmgr"
 
 	"github.com/crypto-power/cryptopower/libwallet/assets/btc"
-	"github.com/crypto-power/cryptopower/libwallet/assets/ltc"
 	sharedW "github.com/crypto-power/cryptopower/libwallet/assets/wallet"
 	"github.com/crypto-power/cryptopower/libwallet/utils"
 )
@@ -156,26 +155,4 @@ func (mgr *AssetsManager) BTCWalletWithSeed(seedMnemonic string) (int, error) {
 		}
 	}
 	return -1, nil
-}
-
-// RestoreLTCWallet restores a LTC wallet from a seed and returns it.
-func (mgr *AssetsManager) RestoreLTCWallet(walletName, seedMnemonic, privatePassphrase string, privatePassphraseType int32) (sharedW.Asset, error) {
-	pass := &sharedW.AuthInfo{
-		Name:            walletName,
-		PrivatePass:     privatePassphrase,
-		PrivatePassType: privatePassphraseType,
-	}
-	wallet, err := ltc.RestoreWallet(seedMnemonic, pass, mgr.params)
-	if err != nil {
-		return nil, err
-	}
-
-	mgr.Assets.LTC.Wallets[wallet.GetWalletID()] = wallet
-
-	// extract the db interface if it hasn't been set already.
-	if mgr.db == nil && wallet != nil {
-		mgr.setDBInterface(wallet.(sharedW.AssetsManagerDB))
-	}
-
-	return wallet, nil
 }
